feat: ツール実装・パネル改善・設定更新
- BrushTool / FillTool / MoveLayerTool: 実装改善
- ToolDescriptor / ToolPanel / SubToolPanel: パネル機能拡張
- ToolTypes: ツール定義更新
- HANDOFF_STATE / STATUS_REPORT: 進行状況更新
- vcpkg.json: 依存パッケージ更新

+++ src/core/tools/FillTool.h
@@ -34,8 +34,8 @@ public:
   const Settings& settings() const noexcept { return m_settings; }
 
 private:
-  static bool isSameColor(const Color& a, const Color& b) noexcept;
-  static int colorDistance(const Color& a, const Color& b) noexcept;
+  static bool  isSameColor(const Color& a, const Color& b) noexcept;
+  static float colorDistance(const Color& a, const Color& b) noexcept;  // 知覚的ユークリッド距離
   bool matchesTarget(const PixelBuffer& source, const Color& target, int x, int y) const noexcept;
   bool hasBridge(const PixelBuffer& source, const Color& target, int x, int y) const noexcept;
